Merge pull request !189 from Shirong Wang/fix-occ-san

+++ src/scf_io/fchk.rs
@@ -74,9 +74,9 @@ impl SCF {
         write!(input, "Number of atoms                            I {:16}\n", self.mol.geom.elem.len());
         write!(input, "Charge                                     I {:16}\n", self.mol.ctrl.charge as i32);
         write!(input, "Multiplicity                               I {:16}\n", self.mol.ctrl.spin as i32);
-        write!(input, "Number of electrons                        I {:16}\n", self.mol.num_elec[0]);
-        write!(input, "Number of alpha electrons                  I {:16}\n", self.mol.num_elec[1]);
-        write!(input, "Number of beta electrons                   I {:16}\n", self.mol.num_elec[2]);
+        write!(input, "Number of electrons                        I {:16}\n", self.mol.num_elec[0] as i32);
+        write!(input, "Number of alpha electrons                  I {:16}\n", self.mol.num_elec[1] as i32);
+        write!(input, "Number of beta electrons                   I {:16}\n", self.mol.num_elec[2] as i32);
         write!(input, "Number of basis functions                  I {:16}\n", self.mol.num_basis);
         write!(input, "Number of independent functions            I {:16}\n", self.mol.num_state);
         // ==============================
